:truck: feat: rename DocBlockHelper to PhpDocHelper
- This was done because there is a new PhpDoc sense and it makes sense to move the class into there.
