+++ rules/Silverstripe52/Rector/Class_/AddManyManyMethodAnnotationsToDataObjectRector.php
@@ -62,7 +62,7 @@ CODE_SAMPLE
             ManyManyList::class
         );
 
-        return $this->docBlockHelper->convertTypesToMethodTagValueNodes(
+        return $this->phpDocHelper->convertTypesToMethodTagValueNodes(
             $manyManyMethods
         );
     }
